test or find a password
- This program tests a given password and checks if it can be easily guessed
- It also finds a particular password from a preexisting file. (ie: common_passwords and dictionary)
- There are two ways the program tries to find a password: brute force if password is short enough, and check from a file of common passwords

**Note: Under Section 342.1 of the Canadian Criminal Code,
it is a federal crime in Canada to have unauthorized access to
password-protected data and usage of said data or password for the
intent to commit an offence may result in severe consequences.
The maximum sentence is 10 years of imprisonment.
The Passwords Program is only used for python coding practice.